feat: star the suggested sample apps
