+++ packages/fx-core/src/common/samples-config.json
@@ -9,7 +9,8 @@
             "fullDescription": "This is a Hello World app of Microsoft Teams Tab app which accomplishes very simple function like single-sign on. You can run this app locally or deploy it to Microsoft Azure. This app only has a Tab frontend.",
             "tags": ["Tab", "TS"],
             "time": "5min to run",
-            "configuration": "Ready for local debug"
+            "configuration": "Ready for local debug",
+            "suggested": false
         },
         {
             "id": "hello-world-tab-with-backend",
@@ -18,7 +19,8 @@
             "fullDescription": "This is a Hello World app of Microsoft Teams Tab app which accomplishes very simple function like single-sign on. You can run this app locally or deploy it to Microsoft Azure. This app has a Tab frontend and a backend service using Azure Function.",
             "tags": ["Tab", "TS", "Azure function"],
             "time": "5min to run",
-            "configuration": "Ready for local debug"
+            "configuration": "Ready for local debug",
+            "suggested": false
         },
         {
             "id": "hello-world-bot",
@@ -27,7 +29,8 @@
             "fullDescription": "This is a Hello World app of Microsoft Teams Bot app demonstrating how to create commands and build adaptive cards that best for automation.",
             "tags": ["Bot", "TS", "Adaptive Cards"],
             "time": "5min to run",
-            "configuration": "Ready for local debug"
+            "configuration": "Ready for local debug",
+            "suggested": false
         },
         {
             "id": "bot-sso",
@@ -41,7 +44,8 @@
                 "SSO"
             ],
             "time": "5min to run",
-            "configuration": "Ready for local debug"
+            "configuration": "Ready for local debug",
+            "suggested": false
         },
         {
             "id": "graph-toolkit-contact-exporter",
@@ -50,7 +54,8 @@
             "fullDescription": "Contact Exporter app can export your teams contact info to Excel files, and it is using Microsoft Graph Toolkit as UI component. This app only has a Tab frontend.",
             "tags": ["Tab", "JS"],
             "time": "5min to run",
-            "configuration": "Ready for local debug"
+            "configuration": "Ready for local debug",
+            "suggested": false
         },
         {
             "id": "todo-list-with-Azure-backend",
@@ -59,7 +64,8 @@
             "fullDescription": "Todo List provides an easy way to manage to-do items in Teams Client. This app helps enabling task collaboration and management for your team. The frontend is a React app and the backend is hosted on Azure. You will need an Azure subscription to run the app.",
             "tags": ["React", "Azure function", "Azure SQL", "JS", "CI/CD"],
             "time": "30min to run",
-            "configuration": "Manual configurations required"
+            "configuration": "Manual configurations required",
+            "suggested": false
         },
         {
             "id": "todo-list-SPFx",
@@ -68,7 +74,8 @@
             "fullDescription": "Todo List with SPFx is a Todo List for individuals to manage his/her personal to-do items. This app is hosted on Sharepoint. There is no requirements to deploy Azure resources.",
             "tags": ["SharePoint", "SPFx", "TS"],
             "time": "1hr to run",
-            "configuration": "Manual configurations required"
+            "configuration": "Manual configurations required",
+            "suggested": false
         },
         {
             "id": "share-now",
@@ -77,7 +84,8 @@
             "fullDescription": "Share Now promotes the exchange of information between colleagues by enabling users to share content within the Teams environment. Users engage the app to share items of interest, discover new shared content, set preferences, and bookmark favorites for later reading.",
             "tags": ["Tab", "Messaging extension", "TS"],
             "time": "30min to run",
-            "configuration": "Ready for local debug"
+            "configuration": "Ready for local debug",
+            "suggested": false
         },
         {
             "id": "in-meeting-app",
@@ -86,7 +94,8 @@
             "fullDescription": "In-meeting app is a hello-world template which shows how to build an app in the context of a Teams meeting. This is a hello-world sample which does not provide any functional feature. This app contains a side panel and a Bot which only shows user profile and can only be added to a Teams meeting.",
             "tags": ["Meeting extension", "JS"],
             "time": "4hr to run",
-            "configuration": "Manual configurations required"
+            "configuration": "Manual configurations required",
+            "suggested": false
         },
         {
             "id": "faq-plus",
@@ -95,7 +104,8 @@
             "fullDescription": "FAQ Plus is a conversational Q&A bot providing an easy way to answer frequently asked questions by users. One can ask a question and the bot responds with information in the knowledge base. If the answer is not in the knowledge base, the bot submits the question to a pre-configured team of experts who help provide support.",
             "tags": ["Easy QnA", "Bot", "JS"],
             "time": "1hr to run",
-            "configuration": "Manual configurations required"
+            "configuration": "Manual configurations required",
+            "suggested": false
         },
         {
             "id": "todo-list-with-Azure-backend-M365",
@@ -104,7 +114,8 @@
             "fullDescription": "Todo List app helps to manage personal to do items and also runs across Microsoft 365, including Teams, Outlook and Office. The frontend is a React app and the backend is an Azure Function. You can optionally deploy and host the app in Azure.",
             "tags": ["React", "Azure function", "JS", "Outlook", "Office"],
             "time": "30min to run",
-            "configuration": "Manual configurations required"
+            "configuration": "Manual configurations required",
+            "suggested": false
         },
         {
             "id": "NPM-search-connector-M365",
@@ -113,7 +124,8 @@
             "fullDescription": "NPM Search Connector is a Messaging Extension that runs across M365 including Teams and Outlook (Web). It allows you to perform a quick search for a NPM package and insert package details into conversations for easy sharing. The front end is built with Adaptive Cards and the backend is an Azure Bot Service.",
             "tags": ["Messaging extension", "JS", "Outlook"],
             "time": "45min to run",
-            "configuration": "Manual configurations required"
+            "configuration": "Manual configurations required",
+            "suggested": false
         }
     ]
 }
